feat(Instances): использовать каскадную вставку
